+++ android/src/main/kotlin/io/flutter/plugins/regula/documentreader/flutter_document_reader_api/Utils.kt
@@ -1,5 +1,5 @@
 //
-//  Utils.java
+//  Utils.kt
 //  DocumentReader
 //
 //  Created by Pavel Masiuk on 21.09.2023.